refactore : create GeneralController for put all dashboard route in function in the controller
